jpocr/requirements.txt
autopep8==2.0.2
black==23.3.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.3
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
mypy-extensions==1.0.0
numpy==1.24.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pycodestyle==2.10.0
PyMuPDF==1.22.2
pyocr==0.7.2
pyparsing==3.0.9
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
typing_extensions==4.5.0
zipp==3.15.0
